+++ kerb_main/C_Tclient.py
@@ -7,12 +7,12 @@ Description:
 import socket as sk
 from MsgFieldDef import *
 
-ID_C = 10  # !每个C的ID需不同
-C_IP = '192.168.137.1'  # !IP需提前声明
+ID_C = 13  # !每个C的ID需不同
+C_IP = '192.168.137.60'  # !IP需提前声明
 
 AS_IP, AS_PORT = '192.168.137.1', 8010
-TGS_IP, TGS_PORT = '192.168.137.1', 8020
-V_IP, V_PORT = '192.168.137.1', 8030
+TGS_IP, TGS_PORT = '192.168.137.51', 8020
+V_IP, V_PORT = '192.168.137.60', 8030
 MAX_SIZE = 2048